move user type into form meta data
